+++ view/proveedores.php
@@ -143,7 +143,7 @@ $proveedores = new Database();
                   <td><?php echo $telefono; ?></td>
                   <td>
                     <a href="../crud/updateProveedor.php?idproveedor=<?php echo $idproveedor; ?>" class="edit" title="Editar" data-toggle="tooltip"><i class="material-icons">&#xE254;</i></a>
-                    <a href="../crud/deleteProveedor.php?idproveedor=<?php echo $idproveedor; ?>" class="delete" title="Eliminar" data-toggle="tooltip"><i class="material-icons">&#xE872;</i></a>
+                    <a href="../crud/deleteProveedor.php?idproveedor=<?php echo $idproveedor; ?>" class="delete" title="Eliminar" data-toggle="tooltip" onclick="return confirm('Seguro que quieres eliminarlo?')"><i class="material-icons">&#xE872;</i></a>
                   </td>
                 </tr>
               <?php
